+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.12.0 (2025-09-23)
 - ✨ Added improved encryption algoriphm `AES-GCM` with WEB Crypto API without breaking backward compatibility [#54](https://github.com/tab-guardian/tab-guardian/issues/54)
 - ✨ Added showing a warning message on private groups that use the old encryption implementation to suggest them to re-encrypt for improved security
+- ✨ Added a loading spinner that appears when you unlock and lock the private group [#55](https://github.com/tab-guardian/tab-guardian/issues/55)
 
 ## v1.11.0 (2025-09-10)
 - 🧑‍💻 Improve helping messages in settings